requirements/test.txt
-r base.txt

pytest-django
pytest-mock